Fix test

+++ tests/test_graph.py
@@ -753,8 +753,8 @@ def _test_add_nodes_sequence(flag = gamera.graph.FREE, count = 2000):
 
    g.add_nodes(list(range(1000,1000+count)))
 
-   assert g.has_node(1000+count/2) == True
-   assert g.get_node(1000+count/2)() == 1000+count/2
+   assert g.has_node(int(1000+count/2)) == True
+   assert g.get_node(int(1000+count/2))() == int(1000+count/2)
 
    for i in range(1000,1000+count):
       assert g.has_node(i) == True
